Работы 3-4.1 com 3

+++ students/23K0563/23K0563-p03/src/main/java/mirea/lab3/Prac3p1p1.java
@@ -16,7 +16,7 @@ public abstract class Prac3p1p1 {
             float[] a = new float[sc.nextInt()];
             for (int i = 0; i < a.length; i++) {
                 if (i % 2 == 0) {
-                    a[i] = rand.nextFloat((float) 100);
+                    a[i] = rand.nextFloat(100);
                 } else {
                     a[i] = (float) Math.random() * 100;
                 }
